+++ plane_above/flight.py
@@ -1,31 +1,16 @@
 import asyncio
-from dataclasses import dataclass
 from urllib.parse import urljoin
 
 import httpx
 
 from .utils import log, async_get
+from .models import Flight, Airport
 from .static import (
     ROUTE_DETAILS_HX_SOURCE_URL,
     ROUTE_DETAILS_SB_SOURCE_URL,
     AIRPORT_DETAILS_AD_SOURCE_URL,
     AIRPORT_DETAILS_HX_SOURCE_URL,
 )
-
-
-@dataclass(frozen=True)
-class Airport:
-    iata: str = "N/A"
-    name: str = "Unknown airport"
-    country_code: str = ""
-
-
-@dataclass(frozen=True)
-class Route:
-    departure: Airport
-    destination: Airport
-    stops: list[Airport]
-    airline: str | None = None
 
 
 class FlightRoute:
@@ -38,12 +23,12 @@ class FlightRoute:
         result = await async_get(_client, AIRPORT_DETAILS_AD_SOURCE_URL, params=dict(iata=iata))
         return Airport(
             iata=iata,
-            name=result.json_data.get("name", "Unknown airport"),
-            country_code=result.json_data.get("country_code") or "",
+            name=result.json_data.get("name") or "Unknown airport",
+            country_code=result.json_data.get("country_code") or Airport.country_code,
         )
 
     @classmethod
-    async def _get_route_from_hx(cls, _client: httpx.AsyncClient, callsign: str) -> Route | None:
+    async def _get_route_from_hx(cls, _client: httpx.AsyncClient, callsign: str) -> Flight | None:
         result = await async_get(_client, urljoin(ROUTE_DETAILS_HX_SOURCE_URL, callsign))
         if (result.status_code != httpx.codes.NOT_FOUND) and (route := result.json_data.get("route")):
             try:
@@ -53,7 +38,7 @@ class FlightRoute:
                     cls._get_airport_details(_client, destination_iata),
                     *[cls._get_airport_details(_client, stop_iata) for stop_iata in stops],
                 )
-                return Route(departure=departure, destination=destination, stops=stops)
+                return Flight(callsign=callsign, departure=departure, destination=destination, stops=stops)
 
             except ValueError:
                 log.error(f" ✈ Cannot parse route for {callsign}: {route}.")
@@ -61,7 +46,7 @@ class FlightRoute:
         return None
 
     @staticmethod
-    async def _get_route_from_sb(_client: httpx.AsyncClient, callsign: str) -> Route | None:
+    async def _get_route_from_sb(_client: httpx.AsyncClient, callsign: str) -> Flight | None:
         result = await async_get(_client, urljoin(ROUTE_DETAILS_SB_SOURCE_URL, callsign))
         if not httpx.codes.is_success(result.status_code) or not result.json_data:
             return None
@@ -70,23 +55,24 @@ class FlightRoute:
         departure = response.get("origin", {})
         destination = response.get("destination", {})
         stop = response.get("midpoint", {})
-        airline = response.get("airline", {}).get("name")
-        return Route(
+        airline = response.get("airline", {}).get("name") or Flight.airline
+        return Flight(
+            callsign=callsign,
             departure=Airport(
-                iata=departure.get("iata_code"),
-                name=departure.get("name"),
-                country_code=departure.get("country_iso_name"),
+                iata=departure.get("iata_code") or Airport.iata,
+                name=departure.get("name") or Airport.name,
+                country_code=departure.get("country_iso_name") or Airport.country_code,
             ),
             destination=Airport(
-                iata=destination.get("iata_code"),
-                name=destination.get("name"),
-                country_code=destination.get("country_iso_name"),
+                iata=destination.get("iata_code") or Airport.iata,
+                name=destination.get("name") or Airport.name,
+                country_code=destination.get("country_iso_name") or Airport.country_code,
             ),
             stops=[
                 Airport(
-                    iata=stop.get("iata_code"),
-                    name=stop.get("name"),
-                    country_code=stop.get("country_iso_name"),
+                    iata=stop.get("iata_code") or Airport.iata,
+                    name=stop.get("name") or Airport.name,
+                    country_code=stop.get("country_iso_name") or Airport.country_code,
                 ),
             ]
             if stop
@@ -95,7 +81,7 @@ class FlightRoute:
         )
 
     @classmethod
-    async def get_route(cls, _client: httpx.AsyncClient, callsign: str) -> Route:
+    async def get_route(cls, _client: httpx.AsyncClient, callsign: str) -> Flight:
         if callsign:
             if route := await cls._get_route_from_hx(_client, callsign):
                 return route
@@ -103,9 +89,10 @@ class FlightRoute:
             if route := await cls._get_route_from_sb(_client, callsign):
                 return route
 
-        return Route(
+        return Flight(
+            callsign=callsign or "",
             departure=Airport(name="Unknown departure"),
             destination=Airport(name="Unknown destination"),
             stops=[],
-            airline=None,
+            airline=Flight.airline,
         )