+++ plane_above/utils/__init__.py
@@ -1,4 +1,5 @@
 from .http import HttpResult, async_get
 from .logger import log
+from .http.client import PlaneAboveClient
 
-__all__ = ["HttpResult", "async_get", "log"]
+__all__ = ["HttpResult", "PlaneAboveClient", "async_get", "log"]