+++ tests/test_osn_filter.py
@@ -1,4 +1,5 @@
-from plane_above.osn import OSN, FlyingObject
+from plane_above.osn import OSN
+from plane_above.models import State, FlyingObject
 
 from .data.osn.flying_objects import (
     UNKNOWN,
@@ -32,7 +33,14 @@ def test_no_country():
 def test_no_callsign():
     filtered, quantity = OSN._filter_objects(NO_CALLSIGN)
     assert quantity == 1
-    assert filtered == [FlyingObject(icao24="44e697", callsign="", country_code="BE", velocity=4, altitude=82)]
+    assert filtered == [
+        FlyingObject(
+            icao24="44e697",
+            callsign="",
+            country="Belgium",
+            state=State(velocity=4, altitude=82, latitude=51.0031, longitude=5.063),
+        )
+    ]
 
 
 def test_zero_altitude():
@@ -42,9 +50,8 @@ def test_zero_altitude():
         FlyingObject(
             icao24="505cbb",
             callsign="OMS575",
-            country_code="SK",
-            velocity=156,
-            altitude=0,
+            country="Slovakia",
+            state=State(velocity=156, altitude=0, latitude=50.4194, longitude=6.2806),
         )
     ]
 
@@ -56,9 +63,8 @@ def test_zero_velocity():
         FlyingObject(
             icao24="39e680",
             callsign="AFR94XK",
-            country_code="FR",
-            velocity=0,
-            altitude=11567,
+            country="France",
+            state=State(velocity=0, altitude=11567, latitude=52.5609, longitude=5.7613),
         )
     ]
 
@@ -70,8 +76,7 @@ def test_country_code_convert():
         FlyingObject(
             icao24="76cec5",
             callsign="SIA318",
-            country_code="UN",
-            velocity=897,
-            altitude=11232,
+            country="S1ngap0re",
+            state=State(velocity=897, altitude=11232, latitude=51.5289, longitude=5.4091),
         )
     ]