+++ CHANGELOG.md
@@ -1,8 +1,51 @@
 # Changelog
 
-### 0.1.1 - 03/08/2026
-- Fix: add explicit timeouts to http calls
-- Docs: update README
+All notable changes to this project will be documented in this file.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/), and this project
+adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### 0.1.0 - 03/08/2026
+## [0.2.0] - 2026-08-11
+
+
+### Added
+
+- `.aircraft.country` _e.g. "Kingdom of the Netherlands"_
+- `.aircraft.type_code` _e.g. "A35K"_
+- `.aircraft.photos` returning list with `Photo` object _(currently limited to 1)_
+- `.state.latitude` and `.state.longitude`
+
+
+### Changed
+
+- `Plane` renamed to `Above`
+- `Route` renamed to `Flight`
+- `.route` renamed to `.flight`
+- `.callsign` moved to `.flight.callsign`
+- `.icao24` moved to `.aircraft.icao24`
+
+
+### Deprecated
+
+> ⚠️ Deprecated attributes below still work but emit `DeprecationWarning` and will be removed in **v1.0.0**.
+- `Plane` and `Route` models
+- `.route`, `.callsign`, `.country_code`, `.icao24` and `.photo` attributes
+
+
+### Fixed
+
+- Lock unavailable sources to prevent hitting timeouts constantly
+- `.photo.image_url` and `.photo.origin_url` might point to different photos in some cases
+
+
+## [0.1.1] - 2026-08-03
+
+### Fixed
+
+- Explicit timeouts in http calls
+
+
+## [0.1.0] - 2026-08-03
+
+### Added
+
 - Initial release